requirements.txt
aiofiles==24.1.0
aiogram==2.15
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
arcade==2.6.17
asgiref==3.8.1
async-timeout==5.0.1
asyncio==3.4.3
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.16.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.1.6
django-cors-headers==4.6.0
django-eav2==1.7.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.114.2
filelock==3.16.0
Flask==3.0.3
flexcache==0.3
flexparser==0.4
frozenlist==1.5.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
gspread==6.1.4
h11==0.14.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
instagrapi==2.1.2
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
MouseInfo==0.1.3
multidict==6.1.0
numpy==2.1.0
oauthlib==3.2.2
openai==0.28.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
Panda3D==1.10.15
panda3d-gltf==1.2.1
panda3d-simplepbr==0.12.0
pandas==2.2.2
pefile==2023.2.7
Pillow==9.3.0
Pint==0.24.4
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.0.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pygame==2.5.2
PyGetWindow==0.0.9
pyglet==2.0.dev23
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyJWT==2.8.0
PyMsgBox==1.0.9
pymunk==6.4.0
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytiled-parser==2.2.0
pytils==0.4.1
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.3
qrcode==8.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
screeninfo==0.8.1
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SpeechRecognition==3.14.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
starlette==0.38.5
tomli==2.1.0
tqdm==4.67.0
trio==0.26.0
trio-websocket==0.11.1
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
ursina==7.0.0
virtualenv==20.26.4
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xlwings==0.32.1
yarl==1.18.3
